Fix uv sync flag in CI workflow
