+++ pyproject.toml
@@ -7,6 +7,23 @@ dependencies = [
     "torch>=2.8.0",
 ]
 
+[project.optional-dependencies]
+dev = [
+    "pre-commit",
+    "pytest",
+    "ruff",
+]
+
+[dependency-groups]
+dev = [
+    "pre-commit",
+    "pytest",
+    "ruff",
+]
+
+[tool.ruff]
+line-length = 88
+
 [[tool.uv.index]]
 url = "https://pypi.org/simple"
